requirements.txt
python-telegram-bot==21.3
google-api-python-client==2.134.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
